Add a new 'Update' mode
TWENTY_TWENTY=overwrite will ALWAYS overwrite the image,
even if the results are within the threshold. This makes
PRs full of noisy file changes that are totally within
the threshold and didn't need to be updated.

The new TWENTY_TWENTY=update mode will only write
the image to disk if the results were beyond the
threshold. This way, PRs only show relevant changes
that need inspecting (i.e. were beyond the threshold)